chs-edge/requirements.txt
numpy
paho-mqtt
pymodbus
python-dotenv
requests
